+++ src/main/java/lv/venta/repo/IMacibuRezultatiRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import lv.venta.model.MacibuRezultati;
 
-public interface IMacibuRezultatiRepo extends CrudRepository<MacibuRezultati, Integer> {
+public interface IMacibuRezultatiRepo extends CrudRepository<MacibuRezultati, Long> {
